# docs/openapi.yaml
openapi: 3.0.3
info:
  title: OpenClaw Context Engine API
  description: |
    OpenClaw 上下文引擎中间层 API —— 为 AI Agent 提供持久记忆和结构化知识检索。
    
    ## 架构
    - **localMem**: SQLite 记忆存储（2-state 模型：tentative / kept）
    - **LLMWiki**: 结构化知识页面编译和搜索
    
    ## 认证
    配置 `OPENCLAW_API_SECRET` 后，所有请求需要 Bearer Token。
    Unix Socket 连接自动注入 Token。
    
    ## 速率限制
    默认 100 请求/分钟，可通过环境变量配置。
  version: 1.0.0
  contact:
    name: OpenClaw Team
  license:
    name: MIT
servers:
  - url: http://127.0.0.1:8901
    description: 本地 HTTP 服务
  - url: http://localhost
    description: Unix Socket 代理（通过 /tmp/openclaw-engine.sock）
security:
  - bearerAuth: []
...
paths:
...
  /api/wiki/check-stale:
    get:
      tags: [Wiki]
      summary: 检查 Wiki 是否需要重新编译
      description: 轻量级检查，不读取文件内容
      responses:
        '200':
          description: 检查成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      stale:
                        type: boolean
                      summary:
                        type: object
        '500':
          $ref: '#/components/responses/InternalError'
  /api/wiki/remove-page:
    post:
      tags: [Wiki]
      summary: 删除 Wiki 页面
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [wikiPageName]
              properties:
                wikiPageName:
                  type: string
      responses:
        '200':
          description: 删除成功
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/wiki/update-index:
    post:
      tags: [Wiki]
      summary: 更新 Wiki 总索引
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                pages:
                  type: array
                  items:
                    type: object
                    properties:
                      name:
                        type: string
                      title:
                        type: string
                      sourceId:
                        type: string
      responses:
        '200':
          description: 更新成功
        '500':
          $ref: '#/components/responses/InternalError'
  /api/wiki/compile-prompt:
    post:
      tags: [Wiki]
      summary: 生成 Wiki 编译提示词
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                changesResult:
                  type: object
      responses:
        '200':
          description: 生成成功
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/wiki/search:
    post:
      tags: [Wiki]
      summary: 搜索 Wiki 页面
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WikiSearchRequest'
      responses:
        '200':
          description: 搜索成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WikiSearchResponse'
        '422':
          description: 查询仅包含元数据
        '500':
          $ref: '#/components/responses/InternalError'
  /api/wiki/detect-changes:
    post:
      tags: [Wiki]
      summary: 检测 Wiki 源文件变更
      responses:
        '200':
          description: 检测成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WikiChangesResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/wiki/save-page:
    post:
      tags: [Wiki]
      summary: 保存 Wiki 页面
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WikiSavePageRequest'
      responses:
        '200':
          description: 保存成功
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/wiki/status:
    get:
      tags: [Wiki]
      summary: 获取 Wiki 编译状态
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API 认证密钥（OPENCLAW_API_SECRET）
...
  schemas:
...
    WikiSearchRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
          minLength: 1
        top_k:
          type: integer
          default: 5
    WikiSavePageRequest:
      type: object
      required: [sourcePath, wikiPageName, content]
      properties:
        sourcePath:
          type: string
        wikiPageName:
          type: string
        content:
          type: string
        sourceId:
          type: string
...
    WikiSearchResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            results:
              type: array
              items:
                type: object
                properties:
                  pageName:
                    type: string
                  score:
                    type: number
                  snippet:
                    type: string
    WikiChangesResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            hasChanges:
              type: boolean
            added:
              type: array
            modified:
              type: array
            deleted:
              type: array
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        message:
          type: string
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
        details:
          type: object
  responses:
    BadRequest:
      description: 请求验证失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: VALIDATION_ERROR
            message: "Validation failed: query is required"
...
    InternalError:
      description: 服务器内部错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: INTERNAL_ERROR
            message: "An unexpected error occurred"
...